developing the agent

+++ one_agent/model.py
@@ -68,6 +68,6 @@ class Critic(nn.Module):
     def forward(self, state, action):
         """Build a critic (value) network that maps (state, action) pairs -> Q-values."""
         xs = F.relu(self.fcs1(state))
-        x = torch.cat((xs, action), dim=1)
+        x = torch.cat((xs, action.float()), dim=1)
         x = F.relu(self.fc2(x))
         return self.fc3(x)
